+++ familyhelper-ui-view/packages/webapp-pc/src/views/nodes/comvisual/miscellaneous/compregFallback/CompregFallback.vue
@@ -2,15 +2,16 @@
   <div class="compreg-fallback-container">
     <border-layout-panel class="border-layout-panel">
       <div class="main-container">
-        <div class="placeholder">
-          当你看到这个页面的时候，意味着 router.component.key 没有正确配置
-        </div>
+        <placeholder-panel
+          text="当你看到这个页面的时候，意味着 router.component.key 没有正确配置"
+        />
       </div>
     </border-layout-panel>
   </div>
 </template>
 
 <script setup lang="ts">
+import PlaceholderPanel from '@/components/comvisual/layout/placeholderPanel/PlaceholderPanel.vue'
 import BorderLayoutPanel from '@/components/elementPlus/layout/borderLayoutPanel/BorderLayoutPanel.vue'
 
 defineOptions({
@@ -27,18 +28,5 @@ defineOptions({
 .main-container {
   width: 100%;
   height: 100%;
-}
-
-.placeholder {
-  width: 100%;
-  height: 100%;
-  text-align: center;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  font-size: 24px;
-  font-weight: bold;
-  color: #bfbfbf;
-  user-select: none;
 }
 </style>